refactor(api): Rename traits package to `trait`
