+++ modules/telestion-api/src/main/java/de/wuespace/telestion/api/verticle/trait/ExtendedMessageHandler.java
@@ -1,4 +1,4 @@
-package de.wuespace.telestion.api.verticle.traits;
+package de.wuespace.telestion.api.verticle.trait;
 
 import de.wuespace.telestion.api.message.JsonMessage;
 import io.vertx.core.eventbus.Message;
