+++ modules/telestion-api/src/main/java/de/wuespace/telestion/api/verticle/trait/MessageHandler.java
@@ -1,4 +1,4 @@
-package de.wuespace.telestion.api.verticle.traits;
+package de.wuespace.telestion.api.verticle.trait;
 
 import de.wuespace.telestion.api.message.JsonMessage;
