+++ modules/telestion-example/src/main/java/de/wuespace/telestion/example/PongVerticle.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import de.wuespace.telestion.api.verticle.TelestionConfiguration;
 import de.wuespace.telestion.api.verticle.TelestionVerticle;
 import de.wuespace.telestion.api.verticle.VerticleDeployer;
-import de.wuespace.telestion.api.verticle.traits.WithEventBus;
+import de.wuespace.telestion.api.verticle.trait.WithEventBus;
 import de.wuespace.telestion.example.messages.SimpleMessage;
 import io.vertx.core.Vertx;
 import io.vertx.core.eventbus.Message;
