+++ modules/telestion-example/src/main/java/de/wuespace/telestion/example/SimpleCommandHandler.java
@@ -3,7 +3,7 @@ package de.wuespace.telestion.example;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.wuespace.telestion.api.verticle.TelestionConfiguration;
 import de.wuespace.telestion.api.verticle.TelestionVerticle;
-import de.wuespace.telestion.api.verticle.traits.WithEventBus;
+import de.wuespace.telestion.api.verticle.trait.WithEventBus;
 import de.wuespace.telestion.example.messages.SimpleCommand;
 import io.vertx.core.eventbus.Message;
