+++ modules/telestion-example/src/main/java/de/wuespace/telestion/example/TestVerticle.java
@@ -3,8 +3,8 @@ package de.wuespace.telestion.example;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.wuespace.telestion.api.verticle.TelestionConfiguration;
 import de.wuespace.telestion.api.verticle.TelestionVerticle;
-import de.wuespace.telestion.api.verticle.traits.WithEventBus;
-import de.wuespace.telestion.api.verticle.traits.WithSharedData;
+import de.wuespace.telestion.api.verticle.trait.WithEventBus;
+import de.wuespace.telestion.api.verticle.trait.WithSharedData;
 import de.wuespace.telestion.example.messages.SimpleMessage;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.eventbus.Message;
